fix sequelize model import
