+++ api/src/models/sequelize/chat.ts
@@ -1,12 +1,10 @@
 import Sequelize from 'sequelize';
 import sequelize from '../../util/database';
 
-const Chat = sequelize.define('chat', {
+export const Chat = sequelize.define('chat', {
   id: {
     type: Sequelize.STRING,
     allowNull: false,
     primaryKey: true,
   },
 });
-
-export default Chat;
