+++ api/src/models/sequelize/follow.ts
@@ -1,5 +1,3 @@
 import sequelize from '../../util/database';
 
-const Follow = sequelize.define('follow');
-
-export default Follow;
+export const Follow = sequelize.define('follow');
