+++ api/src/models/sequelize/index.ts
@@ -1,0 +1,7 @@
+export * from './chat';
+export * from './follow';
+export * from './item';
+export * from './message';
+export * from './user';
+export * from './userChat';
+export * from './userItem';
